feat(review): split SimilarityService
